requirements.txt
pylast
pyyaml
